All CoreLocation / MapKit extensions in one place
